python/requirements.txt
pandas
numpy
pykafka
pyspark
mysql-connector-python